backend/requirements.txt
fastapi
uvicorn
pdfplumber
sentence-transformers
faiss-cpu
python-multipart
pydantic
numpy